SOFBMonApp: support disconnections in SOFBMonOrbit
This allows users of the SlowOrb-Mon PV to know when some BPM is
disconnected. Since we can now report it, we allow the seq program to be
launched when not all PVs are connected.

It was necessary to change the max and min timestamp logic slightly to
allow the disconnected PV to be the first one of the array.