SOFBMonApp: add SOFBMonIntlk implementation.
This implementation requires two versions of the state machine, one with
chars for the flag vectors, and one with ints for the value vectors. The
logic used for updating flags as fast as possible on the first
transition from "all zero" can be used for the value vectors as well -
they are unlikely to be all zero at any point - which allows us to share
the implementation for updating the vectors at 10Hz.

SOFBMonIntlk.db should be a .template and installed as such, but, since
we are following the pattern in this repository of using .db files
(which should be fixed), it has to be defined as a .db file, since
otherwise the IOC build system will use the msi tool to compile the
template [1], and the default definitions for EGU and FTVL will be
applied to the resulting .db file in place of the specified macros.

[1] https://docs.epics-controls.org/en/latest/build-system/specifications.html